Add files via upload
So, What Can We Do?
Since there is no vaccine available right now, the only way to handle the spread is to slow down the transmission. As it can be seen even in the under-estimates and from the actual data around us, the sharply increasing number of cases is bound to overwhelm the medical infrastructure of any nation. So, by slowing down the transmission, we don’t actually stop the spread but keep the transmission and the active cases at any point in time well within the limits of the medical handling capacity. This is what is being referred to as “Flattening The Curve”.